+++ APDS9960.c
@@ -32,7 +32,7 @@ FctERR NONNULL__ APDS9960_Init(const uint8_t idx, const I2C_HandleTypeDef * hi2c
 
 	I2C_PERIPHERAL_SET_DEFAULTS(APDS9960, idx, devAddress);
 
-	err = I2C_slave_init(&APDS9960_hal[idx], hi2c, devAddress, I2C_slave_timeout);
+	err = I2C_slave_init(&APDS9960_hal[idx], hi2c, devAddress, APDS9960_hal[idx].cfg.timeout);
 	if (!err)	{ err = APDS9960_Init_Sequence(&APDS9960[idx]); }
 
 	if (err)	{ I2C_set_enable(&APDS9960_hal[idx], false); }