+++ AT42QT1244.c
@@ -19,7 +19,8 @@
 /****************************************************************/
 
 
-static const I2C_slave_t AT42QT1244_defaults = { { pNull, 0, I2C_slave_timeout, I2C_MEMADD_SIZE_8BIT, I2C_FM }, 0, HAL_OK, true, false };
+//! \note Timeout increased compared to default value, ensuring no timeout occurs when writing whole setup block (111 bytes)
+static const I2C_slave_t AT42QT1244_defaults = { { pNull, 0, 30, I2C_MEMADD_SIZE_8BIT, I2C_FM }, 0, HAL_OK, true, false };
 
 I2C_slave_t AT42QT1244_hal[I2C_AT42QT1244_NB];
 
@@ -35,7 +36,7 @@ FctERR NONNULL__ AT42QT1244_Init(const uint8_t idx, const I2C_HandleTypeDef * hi
 
 	I2C_PERIPHERAL_SET_DEFAULTS(AT42QT1244, idx, devAddress);
 
-	err = I2C_slave_init(&AT42QT1244_hal[idx], hi2c, devAddress, I2C_slave_timeout);
+	err = I2C_slave_init(&AT42QT1244_hal[idx], hi2c, devAddress, AT42QT1244_hal[idx].cfg.timeout);
 	if (!err)	{ err = AT42QT1244_Init_Sequence(&AT42QT1244[idx]); }
 
 	if (err)	{ I2C_set_enable(&AT42QT1244_hal[idx], false); }