+++ MCP4725.c
@@ -28,7 +28,7 @@ FctERR NONNULL__ MCP4725_Init(const uint8_t idx, const I2C_HandleTypeDef * hi2c,
 
 	I2C_PERIPHERAL_SET_DEFAULTS(MCP4725, idx, devAddress);
 
-	err = I2C_slave_init(&MCP4725_hal[idx], hi2c, devAddress, I2C_slave_timeout);
+	err = I2C_slave_init(&MCP4725_hal[idx], hi2c, devAddress, MCP4725_hal[idx].cfg.timeout);
 	if (!err)	{ err = MCP4725_Init_Sequence(&MCP4725[idx]); }
 
 	if (err)	{ I2C_set_enable(&MCP4725_hal[idx], false); }