+++ PCF8523.c
@@ -22,7 +22,7 @@ __WEAK FctERR PCF8523_Init(void)
 {
 	FctERR err;
 
-	err = I2C_slave_init(&PCF8523_hal, I2C_PCF8523, PCF8523_BASE_ADDR, I2C_slave_timeout);
+	err = I2C_slave_init(&PCF8523_hal, I2C_PCF8523, PCF8523_BASE_ADDR, PCF8523_hal.cfg.timeout);
 	if (!err)	{ err = PCF8523_Init_Sequence(); }
 
 	if (err)	{ I2C_set_enable(&PCF8523_hal, false); }