+++ TCS3472.c
@@ -32,7 +32,7 @@ FctERR NONNULL__ TCS3472_Init(const uint8_t idx, const I2C_HandleTypeDef * hi2c,
 
 	I2C_PERIPHERAL_SET_DEFAULTS(TCS3472, idx, devAddress);
 
-	err = I2C_slave_init(&TCS3472_hal[idx], hi2c, devAddress, I2C_slave_timeout);
+	err = I2C_slave_init(&TCS3472_hal[idx], hi2c, devAddress, TCS3472_hal[idx].cfg.timeout);
 	if (!err)	{ err = TCS3472_Init_Sequence(&TCS3472[idx]); }
 
 	if (err)	{ I2C_set_enable(&TCS3472_hal[idx], false); }