+++ TSL2591.c
@@ -28,7 +28,7 @@ FctERR NONNULL__ TSL2591_Init(const uint8_t idx, const I2C_HandleTypeDef * hi2c,
 
 	I2C_PERIPHERAL_SET_DEFAULTS(TSL2591, idx, devAddress);
 
-	err = I2C_slave_init(&TSL2591_hal[idx], hi2c, devAddress, I2C_slave_timeout);
+	err = I2C_slave_init(&TSL2591_hal[idx], hi2c, devAddress, TSL2591_hal[idx].cfg.timeout);
 	if (!err)	{ err = TSL2591_Init_Sequence(&TSL2591[idx]); }
 
 	if (err)	{ I2C_set_enable(&TSL2591_hal[idx], false); }