+++ ReleaseNotes.md
@@ -28,10 +28,12 @@ SOFTWARE.
 * template update (splitted into singleton/multiple device(s))
 * more NONNULL__ checks
 * possibility to set different device base address than default set in header (at project level or in globals.h)
+* init function for peripherals use value in structure for timeout init (in case using a custom timeout)
 * AMG88: added
 * APDS9930: added (+ warning if APDS9960 also compiled, OFFSET union removed as coded as signed 8bits integer)
 * APDS9960: added with fixes and default handler
 * APDS99xx: use of config bits (instead of Status register ones) to trigger proper interrupt(s) clear function
+* AT42QT1244: timeout value increased compared to default one (ensuring enough time to write the whole setup block)
 * AT42QT1244: Fix uAT42QT_REG__KEY_DATA bitfield order
 * AT42QT1244: crc calculation not static (to be able to use it to generate crc for setup block)
 * AT42QT1244: AT42QT1244_Send_Setup to write whole setup block at once with updated CRC, also returned in given function parameter